requirements.txt
python-dotenv
simplekml
tqdm
